+++ src/extensions/views/InstallCustomExtension/components/InstallSectionData/InstallSectionData.tsx
@@ -6,12 +6,12 @@ import { ExtensionInstallFormData, Manifest } from "../../types";
 import { InstallExtensionManifestData } from "../InstallExtensionManifestData";
 
 export const InstallSectionData = ({
-  isManifestLoading,
+  isFetchingManifest,
   manifest,
   lastFetchedManifestUrl,
   control,
 }: {
-  isManifestLoading: boolean;
+  isFetchingManifest: boolean;
   manifest: Manifest | undefined;
   lastFetchedManifestUrl: string | undefined;
   control: Control<ExtensionInstallFormData>;
@@ -21,23 +21,29 @@ export const InstallSectionData = ({
     name: "manifestUrl",
   });
 
-  return isManifestLoading ? (
-    <Box>
-      <Box display="flex" flexDirection="column" gap={6}>
-        <Skeleton height={5} __width="292px" />
-        <Skeleton height={12} __width="292px" />
+  if (isFetchingManifest) {
+    return (
+      <Box>
+        <Box display="flex" flexDirection="column" gap={6}>
+          <Skeleton height={5} __width="292px" />
+          <Skeleton height={12} __width="292px" />
+        </Box>
+        <Box marginTop={16}>
+          <Skeleton height={5} __width="106px" />
+          <Skeleton height={5} marginTop={4} __width="356px" />
+          <Skeleton height={5} marginTop={1.5} __width="356px" />
+        </Box>
+        <Box marginTop={11}>
+          <Skeleton height={5} __width="356px" />
+          <Skeleton height={5} marginTop={1.5} __width="356px" />
+        </Box>
       </Box>
-      <Box marginTop={16}>
-        <Skeleton height={5} __width="106px" />
-        <Skeleton height={5} marginTop={4} __width="356px" />
-        <Skeleton height={5} marginTop={1.5} __width="356px" />
-      </Box>
-      <Box marginTop={11}>
-        <Skeleton height={5} __width="356px" />
-        <Skeleton height={5} marginTop={1.5} __width="356px" />
-      </Box>
-    </Box>
-  ) : manifest && lastFetchedManifestUrl === manifestUrlInputValue ? (
-    <InstallExtensionManifestData manifest={manifest} />
-  ) : null;
+    );
+  }
+
+  if (manifest && lastFetchedManifestUrl === manifestUrlInputValue) {
+    return <InstallExtensionManifestData manifest={manifest} />;
+  }
+
+  return null;
 };
